--- areacomment.js
@@ -1,7 +1,0 @@
-export function sanitizeHtml({htmlString}) {
-    return htmlString
-        .replaceAll("&", "&amp;")
-        .replaceAll("<", "&lt;")
-        .replaceAll(">", "&gt;")
-        .replaceAll('"', "&quot;");
-};
